Homework 4 edits #3
