+++ src/components/DishCounter/use-count.js
@@ -1,0 +1,29 @@
+import { useState } from "react";
+
+export const useCount = () => {
+  const [value, setValue] = useState(0);
+
+    const increase = () => {
+      if (value < 5) {
+        setValue(value + 1);
+      }
+    };
+
+  // const increase = useCallback(() => {
+  //   if (value < 5) {
+  //     setValue((currentValue) => currentValue + 1);
+  //   }
+  // }, []);
+
+  const decrease = () => {
+    if (value > 0) {
+      setValue((prevState) => prevState - 1);
+    }
+  };
+
+  return {
+    value,
+    increase,
+    decrease,
+  };
+};
